chore: split tool_result test module to separate file
tool_result.rs: 1362 → 979 lines (-383)

Extract the #[cfg(test)] block at lines 977-1362 into
tool_result/tests.rs, matching the pattern already used by
other modules (sandbox/src/tests.rs, tui/render/tests.rs etc.).
